EC2 VPC updates for default VPC - EUCA-9613
A cloud property is added for enable/disable of default vpc for new
accounts. When enabled any new accounts will have a default vpc created.

Account attributes will now show VPC as the only supported platform when
an account has a default vpc. Addressing is updated to allocate
addresses in the vpc domain when an account has a default vpc. Security
group management is updated to allow use of security group names with
the default vpc.

Launched instances will use the accounts default vpc when present and
security group names can be used with the default vpc.

EUCA-9613