Avoding manifest download on migration destination
Since VBR already contains the correct size of the image,
downloading of the manifest during blobstore artifact
provisioning strictly is not necessary. So, in the case
of run instance request, the manifest is downloaded more
as a sanity check. In the case of migration, the download
is avoided, since the self-signed URL used at run time
may have expired by the time of the migration.

Fixes EUCA-9887