ELB API refresh and VPC support - EUCA-9832
The LoadBalancing service is updated to support the current actions and
messages for AWS/ELB and support for VPC and default VPC with ELB is
added.

Actions are added for tagging, attributes and security group updates.

ELB API usage for working with a default VPC is similar to the existing
EC2-Classic approach other than the management of security groups. In
a default VPC a security group is created per VPC.

ELB with an explict VPC requires the user specifies subnets rather than
availability zones and works with the users existing (VPC) security
groups.

The duplicate (service separation related) message hierarchies for ELB
are removed. LoadBalancingBackendMessage is now an interface and there
is an empty extension class for each message that implements this
interface.

EUCA-9832