Ceph-RBD integration
Images are never flattened (copy blocks over) on the backing Ceph cluster.
Deleting an image renames the image with a known prefix on the Ceph cluster.
A worker thread scans pools in the Ceph cluster for images that are marked for deletion and cleans them up if they don't have any clones/children.
If the image marked for deletion has clones/children, the worker thread leaves them untouched until all children are also deleted.