EC2 binding update to 2014-10-01 API verson - EUCA-10165
EC2 bindings are added for API versions up to 2014-10-01. Bindings are
added for missing messages and existing messages are updated to fix
issues with missing/incorrect properties with the Query API.

Backend messages are moved from the messages module to the compute
common module and cluster messages are separated out where there was
reuse. The backend messages now extend the front end messages, removing
the duplicate compute message hieararchy.

The cluster binding declaration is re-arranged to allow unit testing for
the core message bindings.

EUCA-10165