Add euca-install-service-image, service-images.yml
This commit adds a new "euca-install-service-image" command that is
meant to tell euca2ools to find an image capable of running a certain
service and install it.  This proof-of-concept implementation is much
"dumber" and hardcodes most of that logic in its own code and in the
file "service-images.yml".

A complete implementation of the unified service image tool feature will
require euca2ools to have a full-featured euimage implementation that
includes a dependency solver.  Ideally, the solution could also learn or
infer property names, service names, and service worker API versions
from info it obtains from the CLC, but that can also remain on the
filesystem if that is too complex or time-consuming to implement.

Fixes EUCA-10198, EUCA-10201, and EUCA-10202