Configurable certificate credential download strategy - EUCA-10273
Administrators can now control the strategy for certificate
credential generation on credential download. It is possible to disable
inclusion of a certificate, to generate a certificate only if none
exists or to respect the configured IAM limit for signing certificates.

The limits for access keys and signing certificates are increased to
1,000,000 on upgraded systems so they are effectively disabled.

EUCA-10273