EC2 describe regions support for region federation - EUCA-10450
DescribeRegions now includes compute services from the region
configuration. If there are multiple endpoints defined the first is
used. When regions are not configured we fallback to the old behaviour
except we no longer check the service status (i.e. the local region is
always present in the response)

EUCA-10450