DNS component AWS name and aliases support - EUCA-10907
Components can now be resolved by using their AWS endpoint host label
with the internal and external subdomains in addition to the component
host label. The new ServiceNames annotation can be placed on ComponentId
classes if more explicit specification of aliases is required. This is
now used to restore support for a "eucalyptus" compute service alias.

EUCA-10907