EC2 shutdown timeout and subnet free addresses - EUCA-11363
Nodes without service tags are now shutdown after the shutdown timeout
expires. It is expected that instances in the shutting down state have a
service tag, so we now also check that the tag is not set to null in a
state update.

We now retry the subnet free address calculation if it fails due to an
optimistic locking failure. A check is added to ensure the free address
count is not negative.

EUCA-11363