Suggest installing eucalyptus when rsync exits w/12
If one happens to point clcadmin-copy-keys at a machine that does not
yet have eucalyptus installed on it then /var/lib/eucalyptus/keys will
not exist, making rsync bail out.  This commit adds a check for the
return code that rsync uses when that happens and explicitly mentions
that that may be the case.

Fixes EUCA-11437