Fix for EUCA-11328
An EIP association is implemented in midonet with SNAT and DNAT rules.
DNAT is implemented as a infilter of the VPC router, while SNAT is implemented as an
outfilter of the VPC router.
SNAT rule was only excluding the subnet+2 address (for metadata networking). To address this issue,
the subnet+2 address(es) is/are added to the private ip-address-group of the SG. SNAT rule should
match the SG's private ip-address-group as source and do not perform SNAT.
The obvious implementation is to perform an inverse ip-address-group match. However, there is a bug
in midonet-cli (Midokura ticket MN-3980) that does not parse/process inverse ip-address-group
matches properly.
midolman seems to implement the rule correctly. Due to inconvenience in debugging (once an inverse
ip-address-group matching rule is placed in a chain, listing rules in that chain fails), the temporary
solution is to use a set of 2 rules (one to match the SG's private ip-address-group and accept, and
another to perform SNAT)
In order to enable VM-VM communication via public IPs, early SNAT is performed to packets with destination
matching public IPs of source VM's SG.